+++ components/bridge/api/src/main/java/com/flipperdevices/bridge/api/manager/service/RestartRPCApi.kt
@@ -1,0 +1,5 @@
+package com.flipperdevices.bridge.api.manager.service
+
+interface RestartRPCApi {
+    fun restartRpc()
+}