+++ components/bridge/service/impl/src/main/java/com/flipperdevices/bridge/service/impl/FlipperServiceApiImpl.kt
@@ -150,4 +150,8 @@ class FlipperServiceApiImpl(
         info { "Disconnect successful, close manager" }
         bleManager.close()
     }
+
+    override fun restartRPC() {
+        bleManager.restartRPCApi.restartRpc()
+    }
 }