+++ components/settings/impl/src/main/java/com/flipperdevices/settings/impl/composable/category/DebugCategory.kt
@@ -73,6 +73,11 @@ fun DebugCategory(
                     state = settings.ignoreSubghzProvisioningOnZeroRegion,
                     onSwitchState = debugViewModel::onSwitchIgnoreSubGhzProvisioning
                 )
+                GrayDivider()
+                ClickableElement(
+                    titleId = R.string.debug_restart_rpc,
+                    onClick = debugViewModel::restartRpc
+                )
             }
         }
     }