+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - [Feature] Add support for NFC Shadow Files
 - [Feature] Offer update, if region changes or region file not exit on Flipper
 - [Feature] Do not provide SubGhz provisioning on zero hardware region
+- [Feature] Restart RPC on fail RPC process response or flipper lags
 - [Feature] Prepare for Fdroid app store
 - [FIX] Now search for devices by Mac address and by flipper name
 - [FIX] Subghz provisioning while geoip is null